+++ pkg/controllers/user/controllers.go
@@ -5,8 +5,6 @@ import (
 
 	"github.com/rancher/rancher/pkg/controllers/management/compose/common"
 	"github.com/rancher/rancher/pkg/controllers/user/alert"
-	"github.com/rancher/rancher/pkg/controllers/user/authz"
-	"github.com/rancher/rancher/pkg/controllers/user/authz/podsecuritypolicy"
 	"github.com/rancher/rancher/pkg/controllers/user/dnsrecord"
 	"github.com/rancher/rancher/pkg/controllers/user/endpoints"
 	"github.com/rancher/rancher/pkg/controllers/user/externalservice"
@@ -20,6 +18,8 @@ import (
 	"github.com/rancher/rancher/pkg/controllers/user/nodesyncer"
 	"github.com/rancher/rancher/pkg/controllers/user/nslabels"
 	"github.com/rancher/rancher/pkg/controllers/user/pipeline"
+	"github.com/rancher/rancher/pkg/controllers/user/rbac"
+	"github.com/rancher/rancher/pkg/controllers/user/rbac/podsecuritypolicy"
 	"github.com/rancher/rancher/pkg/controllers/user/secret"
 	"github.com/rancher/rancher/pkg/controllers/user/targetworkloadservice"
 	"github.com/rancher/rancher/pkg/controllers/user/usercompose"
@@ -29,7 +29,7 @@ import (
 
 func Register(ctx context.Context, cluster *config.UserContext, kubeConfigGetter common.KubeConfigGetter, clusterManager healthsyncer.ClusterControllerLifecycle) error {
 	alert.Register(ctx, cluster)
-	authz.Register(cluster)
+	rbac.Register(cluster)
 	healthsyncer.Register(ctx, cluster, clusterManager)
 	helm.Register(cluster, kubeConfigGetter)
 	logging.Register(cluster)
