+++ pkg/controllers/user/rbac/cluster_handler.go
@@ -1,4 +1,4 @@
-package authz
+package rbac
 
 import (
 	"reflect"
