+++ pkg/controllers/user/rbac/namespace_handler.go
@@ -1,4 +1,4 @@
-package authz
+package rbac
 
 import (
 	"encoding/json"
