+++ pkg/controllers/user/rbac/podsecuritypolicy/template.go
