+++ pkg/controllers/user/rbac/project_handler.go
@@ -1,4 +1,4 @@
-package authz
+package rbac
 
 import (
 	"fmt"
