+++ pkg/controllers/user/rbac/prtb_handler.go
@@ -1,4 +1,4 @@
-package authz
+package rbac
 
 import (
 	"fmt"
